# [FIX] mrp: speedup _post_inventory for MO batches