Rewrite _post_inventory to make it work better
with mo batches. This mainly improves the
time it takes to mark_as_done manufacturing
orders from the Manufacturing Orders ListView.


#### speedup

A global speedup is difficult to establish as it depends on the installation or not
of the stock_account module, the inventory valuation of the product category, etc...

Here, the customer DB has 658 MO to mark as done, the stock_account module is installed.

|  Batch Size | Before PR | After PR |
|:-----------:|:----------:|:---------:|
| 10 | 18s | 10s |
| 50 | 81s | 34s |
| 100 | 160s | 65s |
| 250 | 432s | 162s |
| 658 | 1420s | 434s|

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr