@pavelsodomka 2022-02-15
This is our ticket.
We regularly produce batches of 5 000 items. Planning to increase to 20 000 items. Can you please calculate how long the MO confirm will take ? Thanks